requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.3.0
pydantic==2.7.1
streamlit==1.35.0
requests==2.32.2
folium==0.16.0
streamlit-folium==0.20.0
pandas==2.2.2
plotly==5.22.0